server/channel_service/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==4.2.21
django-cors-headers==4.7.0
django-mongoengine==0.5.6
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
mongoengine==0.29.1
msgpack==1.1.0
pika==1.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pymongo==4.12.1
pyOpenSSL==25.0.0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==79.0.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
